feat: HW3 opengl

+++ HW3/lib/dda.hxx
@@ -8,5 +8,5 @@
  * @param titik_akhir vector berisi 2 nilai posisi (x, y)
  * @returns vector berisi nilai-nilai posisi setiap titik
  */
-std::vector<double> dda(std::vector<double> titik_awal,
-                        std::vector<double> titik_akhir);
+std::vector<float> dda(std::vector<float> titik_awal,
+                        std::vector<float> titik_akhir);
